Add address labels
Bitmessage addresses found in incoming messages get a label enclosed in
brackets. For example: "BM-2DAV89w336ovy6BUJnfVRD5B9qipFbRgmr" might become
"[General Chan, BM-2DAV89w336ovy6BUJnfVRD5B9qipFbRgmr]"

The labels are removed on outgoing messages, so that other people don't
see your labels.
"[General Chan, BM-2DAV89w336ovy6BUJnfVRD5B9qipFbRgmr]" would get turned
back into "BM-2DAV89w336ovy6BUJnfVRD5B9qipFbRgmr"

Caveat: If the message gets corrupted, then the label might still leak.
Example:
Incoming message:
"My address is BM-2DAV89w336ovy6BUJnfVRD5B9qipFbRgmr"

Gets turned into:
"My address is [General Chan, BM-2DAV89w336ovy6BUJnfVRD5B9qipFbRgmr]"

When you reply something goes wrong and it gets corrupted:
"My addreseneral Chan, BM-2DAV89w336ovy6BUJnfVRD5B9qipFbRgmr]"

That doesn't get recognized as a label and so doesn't get removed.

If the address is one of the user's own identities then the label
associated with it is used.

If an address is unrecognized then it gets a name based on the hash of the
address. This is makes it easier for readers to quickly identify which
addresses are the same.

Note: This only uses the first two bytes of the hash, one to choose the
first name and one to choose the last name. This is anything but collision
proof. The chance that two randomly selected given addresses will have the
same name is 1/65536. It would be very easy for someone to create an
address with a chosen name. This is only meant for convenience and not
for security. If it is important to you to be certain that an address
matches, then check the actual address.

Thanks to .dok for the address listing code.